feat: Add SqlInspector
feat: Add ability to run on leaked ChatGPT model

The SqlInspector gives us the ability to do pseudo-runtime introspection on SQL queries. It uses sqlglot for parsing, but then builds up a similar AST tree in the correct execution order. As tables, columns, and aliases are added, it keeps track of the current list of tables and columns, making it easy to retrieve the final list of tables and columns that were accessed.

Provides:

- List of touched tables, columns
- Also list of touched string values (from where conditions `name="Ryan"`)
- Raises exceptions if a table or column is used that isn't in the schema. (Works through aliases as well)
- Basic foundation for future runtime introspection tools
- Currently not handled: exclude, union, partition, and probably more...